ramp: fix warning for integer ramps
This fixes the following warning:

| ../../../git/std_blocks/ramp/ramp.c: In function 'ramp_init':
| ../../../git/std_blocks/ramp/ramp.c:47:25: error: using floating point absolute value function 'fabs' when argument is of integer type 'uint16_t' {aka 'short unsigned int'} [-Werror=absolute-value]
|    47 |  inf->slope = ((double) fabs(inf->slope) > 10e-6) ? inf->slope : 1;
|       |                         ^~~~

+++ std_blocks/ramp/ramp.c
@@ -44,7 +44,7 @@ int ramp_init(ubx_block_t *b)
 		goto out;
 
 	inf->slope = (len > 0) ? *val : 1;
-	inf->slope = (fabs(inf->slope) > 10e-6) ? inf->slope : 1;
+	inf->slope = (fabs((double) inf->slope) > 10e-6) ? inf->slope : 1;
 
 	ret=0;
 out:
